proxy: fix subrctx start flow
was originally reusing the callback code when _starting_ a subrctx
request, but it needs to be slightly different. It's possible for a
subrctx to immediately return with an error, which could try to complete
a parent coroutine while it was otherwise running.

instead we ensure an unexpected/early return for a subrctx submits an IO
object to be processed _later_, and we finish processing the main rctx's
queues first.