core: tagged listener test basics
wish I wrote at least this much back with the tagged feature... classic
dumbassery to use shell characters for the tag separator.

This was making the perl exec spawn timedrun via a shell call, which no
longer propagates the SIGINT at the end of the test, which prevented the
tests from stopping.

So now... tag_blah_: is supported as well as [], but will probably
deprecate [].